src-tauri/plugins/assistant_api/pyproject.toml
[tool.poetry]
name = "assistant-api"
version = "0.1.0"
description = ""
authors = ["Josh <joshscina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.7.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
